Author: wessels & Christos Tsantilas Import squid3-largeobj branch, fixing proxying and caching of >2GB objects
this patch converts all variables referring to object sizes & offsets to
64-bits, and gets rid of a lot of related casts.

The cache format is also updated to use 64-bit offsets, using a format
compatible with Squid-2.6.

+++ src/AccessLogEntry.h
@@ -1,6 +1,6 @@
 
 /*
- * $Id: AccessLogEntry.h,v 1.5 2006/05/29 00:14:59 robertc Exp $
+ * $Id: AccessLogEntry.h,v 1.6 2007/08/13 17:20:51 hno Exp $
  *
  *
  * SQUID Web Proxy Cache          http://www.squid-cache.org/
@@ -95,9 +95,9 @@ public:
         }
 
         struct IN_ADDR caddr;
-        size_t size;
-        off_t highOffset;
-        size_t objectSize;
+        int64_t size;
+        int64_t highOffset;
+        int64_t objectSize;
         log_type code;
         int msec;
         const char *rfc931;
